+++ tests/artifactory_test.py
@@ -1,4 +1,5 @@
-from unittest.mock import patch, ANY
+import tempfile
+from unittest.mock import patch
 from pytest import fixture
 
 from release.steps.ReleaseRequest import ReleaseRequest
@@ -10,10 +11,11 @@ from release.utils.buildinfo import BuildInfo
 def release_request():
     return ReleaseRequest('org', 'project', 'version', 'buildnumber', 'branch', 'sha')
 
+
 @fixture
 def buildinfo_multi():
     return  BuildInfo({
-        "buildInfo":{
+        "buildInfo": {
             "statuses": [{
             "status": "it-passed",
             "timestamp": "2022-09-09T14:50:41.770+0000",
@@ -23,10 +25,11 @@ def buildinfo_multi():
         }
     })    
 
+
 @fixture
 def buildinfo():
     return  BuildInfo({
-        "buildInfo":{
+        "buildInfo": {
             "statuses": [{
             "status": "it-passed",
             "repository": "sonarsource-public-builds",
@@ -44,7 +47,8 @@ class RepoxResponse:
         self.text = "{ 'message' : 'done' }"
         self.ok = True
     def json(self):
-        return { 'message' : 'done' }
+        return {'message': 'done'}
+
 
 def test_notify(release_request):
     with patch('release.utils.artifactory.requests.get', return_value=RepoxResponse(200)) as request:
@@ -53,6 +57,7 @@ def test_notify(release_request):
             f"{Artifactory.url}/api/build/{release_request.project}/{release_request.buildnumber}",
             headers={'content-type': 'application/json', 'Authorization': 'Bearer token'}
         )
+
 
 def test_promote(release_request,buildinfo):    
     with patch('release.utils.artifactory.requests.post', return_value=RepoxResponse(200)) as request:
@@ -89,6 +94,7 @@ def test_multi_promote(release_request,buildinfo_multi):
             headers={'content-type': 'application/json', 'Authorization': 'Bearer token'}
         )
 
+
 def test_multi_promote_revoke(release_request,buildinfo_multi):
     with patch('release.utils.artifactory.requests.get', return_value=RepoxResponse(200)) as request:        
         Artifactory("token").promote(release_request,buildinfo_multi,True)
@@ -104,10 +110,11 @@ def test_multi_promote_revoke(release_request,buildinfo_multi):
             headers={'content-type': 'application/json', 'Authorization': 'Bearer token'}
         )
 
+
 def test_download():
     with patch('release.utils.artifactory.urllib.request.urlretrieve') as request:        
         Artifactory("token").download('repo','gid','aid','qual','ext','version')
         request.assert_called_once_with(
             f"{Artifactory.url}/repo/gid/aid/version/aid-version-qual.ext", 
-            '/tmp/aid-version-qual.ext'
+            f'{tempfile.gettempdir()}/aid-version-qual.ext'
         )
